# colexec: improve unordered distinct