--- a/pkg/sql/colexec/hashjoiner.go
+++ b/pkg/sql/colexec/hashjoiner.go
@@ -257,10 +257,6 @@ func (hj *hashJoiner) Init() {
 		hj.hashTableInitialNumBuckets,
 		hj.spec.right.sourceTypes,
 		hj.spec.right.eqCols,
-		// Store all columns from the right source since we need to be able to
-		// export the full batches when falling back to the external hash
-		// joiner.
-		nil, /* colsToStore */
 		allowNullEquality,
 		hashTableFullBuildMode,
 		probeMode,