This commit significantly simplifies and optimizes the unordered
distinct operator by making an observation that during `distinctBuild`
method of the hash table the input batch is updated in-place to include
only the distinct tuples. This allows us to simply return the batch
which avoids an unnecessary copy into the output batch. Additionally,
this commit adjusts the hash table instantiation to only store the
distinct columns.

Release note: None