@yuzefovich 2020-12-07
Benchmarks are [here](https://gist.github.com/yuzefovich/1e5daf83c2af1cc800bf03631d9827b4).

Note that there appears to be a large regression, but rerunning it I got
```
name                                                                                         old time/op    new time/op    delta
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=2/ordCols=0-24    7.03ms ± 0%    8.48ms ± 0%  +20.59%  (p=0.000 n=10+9)
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=4/ordCols=0-24    17.2ms ± 0%    15.1ms ± 0%  -12.08%  (p=0.000 n=9+10)

name                                                                                         old speed      new speed      delta
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=2/ordCols=0-24   596MB/s ± 0%   495MB/s ± 0%  -17.08%  (p=0.000 n=10+9)
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=4/ordCols=0-24   489MB/s ± 0%   556MB/s ± 0%  +13.75%  (p=0.000 n=9+10)

name                                                                                         old alloc/op   new alloc/op   delta
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=2/ordCols=0-24     363kB ± 0%     310kB ± 0%  -14.52%  (p=0.002 n=8+10)
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=4/ordCols=0-24     634kB ± 0%     525kB ± 0%  -17.27%  (p=0.002 n=8+10)

name                                                                                         old allocs/op  new allocs/op  delta
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=2/ordCols=0-24     3.08k ± 0%     2.98k ± 0%   -3.34%  (p=0.000 n=10+10)
Distinct/Unordered/hasNulls=false/newTupleProbability=0.001/rows=262144/cols=4/ordCols=0-24     6.05k ± 0%     5.85k ± 0%   -3.34%  (p=0.000 n=10+10)
```
which looks more reasonable.

I think this change should be beneficial in all cases, and the benchmarks are like to be prone to some variability because of "probability of distinct tuple" and whatnot.